refactoring #2
