+++ kafka-connect-http-connectors/src/main/java/io/github/clescot/kafka/connect/Client.java
@@ -1,23 +1,18 @@
 package io.github.clescot.kafka.connect;
 
 import dev.failsafe.RateLimiter;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.Optional;
 
 public interface Client<E> {
-    Logger LOGGER = LoggerFactory.getLogger(Client.class);
 
 
     String getEngineId();
 
     void setRateLimiter(RateLimiter<E> rateLimiter);
 
-
     Optional<RateLimiter<E>> getRateLimiter();
 
-
-
+    String getPermitsPerExecution();
 
 }
